Added CMake file
Adding valgrind to check the memory leak. Removed the extended warning by the compiler.

+++ main.c
@@ -2,7 +2,7 @@
 
 void bytes_reverser(char* first_byte, size_t total_bytes){
     char temp;
-    for(int i=0; i<total_bytes/2; i++){
+    for(size_t i=0; i<total_bytes/2; i++){
         temp = first_byte[i];
         first_byte[i] = first_byte[total_bytes - i -1];
         first_byte[total_bytes - i - 1] = temp;
